refactor(Menu): Name, Price 래핑 클래스 추가
- Name, Price 패키지 이동

+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -28,7 +28,7 @@ public class MenuRequest {
     }
 
     public Menu toEntity() {
-        return Menu.of(name, BigDecimal.valueOf(price), menuGroupId, menuProducts.stream()
+        return Menu.of(name, price, menuGroupId, menuProducts.stream()
             .map(MenuProductRequest::toEntity)
             .collect(toList()));
     }
